split codes into different files with clearer interfaces and structures
